randomized weather data
