+++ public/index.html
@@ -25,7 +25,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>RoboDomo React</title>
+    <title>RoboDomo Client</title>
     <style>
       :focus {
         outline: 0;
