+++ src/dev/mqtt-sim.js
@@ -1,6 +1,7 @@
-const onOrOff = () => (Math.random() < 0.5 ? "off" : "on");
-
-const randomTemp = () => Math.floor(Math.random() * 60 + 60);
+import { onOrOff, randomTemp, randomPositive } from "./utils";
+import { getHourly } from "./weather/hourly";
+import { getForecast } from "./weather/forecast";
+import { getObservation } from "./weather/observation";
 
 const allMessages = [
   ["TOPIC_ROOT/appletv-theater/status/album", ""],
@@ -16128,1498 +16129,13 @@ const allMessages = [
   ],
   ["weather/92109/status/city", "San Diego"],
   ["weather/92109/status/display_city", "San Diego, CA"],
-  [
-    "weather/92109/status/forecast",
-    [
-      {
-        daylight: "D",
-        daySegment: "Afternoon",
-        description: "More clouds than sun. Cool.",
-        skyInfo: 13,
-        skyDescription: "More clouds than sun",
-        temperature: 69,
-        temperatureDesc: "Cool",
-        comfort: 68.72,
-        humidity: 68,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 10,
-        windDirection: 283,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 3,
-        beaufortDescription: "Gentle breeze",
-        visibility: 8.82,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561402800,
-      },
-      {
-        daylight: "D",
-        daySegment: "Evening",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 63.5,
-        humidity: 80,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 290,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.84,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561424400,
-      },
-      {
-        daylight: "N",
-        daySegment: "Night",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 62.96,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 283,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.74,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561446000,
-      },
-      {
-        daylight: "D",
-        daySegment: "Morning",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 64.4,
-        humidity: 78,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 243,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.81,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561467600,
-      },
-      {
-        daylight: "D",
-        daySegment: "Afternoon",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 68,
-        temperatureDesc: "Cool",
-        comfort: 67.64,
-        humidity: 68,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 9,
-        windDirection: 247,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 3,
-        beaufortDescription: "Gentle breeze",
-        visibility: 8.23,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561489200,
-      },
-      {
-        daylight: "D",
-        daySegment: "Evening",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 63.5,
-        humidity: 79,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 280,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.84,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561510800,
-      },
-      {
-        daylight: "N",
-        daySegment: "Night",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 62,
-        temperatureDesc: "Cool",
-        comfort: 62.24,
-        humidity: 82,
-        dewPoint: 56,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 5,
-        windDirection: 287,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.97,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561532400,
-      },
-      {
-        daylight: "D",
-        daySegment: "Morning",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 65,
-        temperatureDesc: "Cool",
-        comfort: 64.58,
-        humidity: 76,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 262,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.89,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 4,
-        weekday: "Wednesday",
-        utcTime: 1561554000,
-      },
-      {
-        daylight: "D",
-        daySegment: "Afternoon",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 68,
-        temperatureDesc: "Cool",
-        comfort: 67.82,
-        humidity: 67,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 10,
-        windDirection: 260,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 3,
-        beaufortDescription: "Gentle breeze",
-        visibility: 7.75,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 4,
-        weekday: "Wednesday",
-        utcTime: 1561575600,
-      },
-      {
-        daylight: "N",
-        daySegment: "Evening",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 62.78,
-        humidity: 80,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 290,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 7.95,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 4,
-        weekday: "Wednesday",
-        utcTime: 1561597200,
-      },
-      {
-        daylight: "N",
-        daySegment: "Night",
-        description: "Cloudy. Cool.",
-        skyInfo: 17,
-        skyDescription: "Cloudy",
-        temperature: 60,
-        temperatureDesc: "Cool",
-        comfort: 60.44,
-        humidity: 84,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 293,
-        windDesc: "Northwest",
-        windDescShort: "NW",
-        beaufortScale: 1,
-        beaufortDescription: "Light air",
-        visibility: 0,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 4,
-        weekday: "Wednesday",
-        utcTime: 1561618800,
-      },
-      {
-        daylight: "D",
-        daySegment: "Morning",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 65,
-        temperatureDesc: "Cool",
-        comfort: 65.12,
-        humidity: 73,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 237,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 5,
-        weekday: "Thursday",
-        utcTime: 1561640400,
-      },
-      {
-        daylight: "D",
-        daySegment: "Afternoon",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 71,
-        temperatureDesc: "Cool",
-        comfort: 75.74,
-        humidity: 62,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 10,
-        windDirection: 235,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 3,
-        beaufortDescription: "Gentle breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 5,
-        weekday: "Thursday",
-        utcTime: 1561662000,
-      },
-      {
-        daylight: "N",
-        daySegment: "Evening",
-        description: "Cloudy. Cool.",
-        skyInfo: 17,
-        skyDescription: "Cloudy",
-        temperature: 65,
-        temperatureDesc: "Cool",
-        comfort: 65.48,
-        humidity: 78,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 6,
-        windDirection: 286,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 5,
-        weekday: "Thursday",
-        utcTime: 1561683600,
-      },
-      {
-        daylight: "N",
-        daySegment: "Night",
-        description: "Cloudy. Cool.",
-        skyInfo: 17,
-        skyDescription: "Cloudy",
-        temperature: 62,
-        temperatureDesc: "Cool",
-        comfort: 62.78,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 5,
-        windDirection: 247,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 5,
-        weekday: "Thursday",
-        utcTime: 1561705200,
-      },
-      {
-        daylight: "D",
-        daySegment: "Morning",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 68,
-        temperatureDesc: "Cool",
-        comfort: 67.82,
-        humidity: 67,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 226,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 6,
-        weekday: "Friday",
-        utcTime: 1561726800,
-      },
-      {
-        daylight: "D",
-        daySegment: "Afternoon",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 71,
-        temperatureDesc: "Cool",
-        comfort: 76.23,
-        humidity: 59,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 10,
-        windDirection: 255,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 3,
-        beaufortDescription: "Gentle breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 6,
-        weekday: "Friday",
-        utcTime: 1561748400,
-      },
-      {
-        daylight: "N",
-        daySegment: "Evening",
-        description: "Partly cloudy. Cool.",
-        skyInfo: 10,
-        skyDescription: "Partly cloudy",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 64.4,
-        humidity: 77,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 6,
-        windDirection: 260,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "15",
-        iconName: "night_partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 6,
-        weekday: "Friday",
-        utcTime: 1561770000,
-      },
-      {
-        daylight: "N",
-        daySegment: "Night",
-        description: "Partly cloudy. Cool.",
-        skyInfo: 10,
-        skyDescription: "Partly cloudy",
-        temperature: 61,
-        temperatureDesc: "Cool",
-        comfort: 61.88,
-        humidity: 83,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 4,
-        windDirection: 203,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "15",
-        iconName: "night_partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 6,
-        weekday: "Friday",
-        utcTime: 1561791600,
-      },
-      {
-        daylight: "D",
-        daySegment: "Morning",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 69,
-        temperatureDesc: "Cool",
-        comfort: 68.72,
-        humidity: 66,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 277,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 7,
-        weekday: "Saturday",
-        utcTime: 1561813200,
-      },
-      {
-        daylight: "D",
-        daySegment: "Afternoon",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 74,
-        temperatureDesc: "Cool",
-        comfort: 76.87,
-        humidity: 56,
-        dewPoint: 56,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 10,
-        windDirection: 272,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 3,
-        beaufortDescription: "Gentle breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 7,
-        weekday: "Saturday",
-        utcTime: 1561834800,
-      },
-      {
-        daylight: "N",
-        daySegment: "Evening",
-        description: "Partly cloudy. Cool.",
-        skyInfo: 10,
-        skyDescription: "Partly cloudy",
-        temperature: 66,
-        temperatureDesc: "Cool",
-        comfort: 66.38,
-        humidity: 76,
-        dewPoint: 59,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 6,
-        windDirection: 280,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "15",
-        iconName: "night_partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 7,
-        weekday: "Saturday",
-        utcTime: 1561856400,
-      },
-      {
-        daylight: "N",
-        daySegment: "Night",
-        description: "Cloudy. Cool.",
-        skyInfo: 17,
-        skyDescription: "Cloudy",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 64.04,
-        humidity: 80,
-        dewPoint: 58,
-        precipitationProbability: 2,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 5,
-        windDirection: 281,
-        windDesc: "West",
-        windDescShort: "W",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 7,
-        weekday: "Saturday",
-        utcTime: 1561878000,
-      },
-      {
-        daylight: "D",
-        daySegment: "Morning",
-        description: "a mixture of sun and clouds. Cool.",
-        skyInfo: 11,
-        skyDescription: "a mixture of sun and clouds",
-        temperature: 70,
-        temperatureDesc: "Cool",
-        comfort: 75.58,
-        humidity: 62,
-        dewPoint: 57,
-        precipitationProbability: 4,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "29",
-        airDescription: "Comfortable",
-        windSpeed: 6,
-        windDirection: 315,
-        windDesc: "Northwest",
-        windDescShort: "NW",
-        beaufortScale: 2,
-        beaufortDescription: "Light breeze",
-        visibility: 0,
-        icon: "4",
-        iconName: "partly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
-        dayOfWeek: 1,
-        weekday: "Sunday",
-        utcTime: 1561899600,
-      },
-    ],
-  ],
-  [
-    "weather/92109/status/hourly",
-    [
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 68,
-        temperatureDesc: "Cool",
-        comfort: 68.18,
-        humidity: 71,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 8,
-        windDirection: 292,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.79,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561399200,
-        localTime: "1106242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 69,
-        temperatureDesc: "Cool",
-        comfort: 68.54,
-        humidity: 69,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 9,
-        windDirection: 288,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.49,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561402800,
-        localTime: "1206242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "More clouds than sun. Cool.",
-        skyInfo: 13,
-        skyDescription: "More clouds than sun",
-        temperature: 69,
-        temperatureDesc: "Cool",
-        comfort: 68.72,
-        humidity: 69,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 10,
-        windDirection: 285,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.79,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561406400,
-        localTime: "1306242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "More clouds than sun. Cool.",
-        skyInfo: 13,
-        skyDescription: "More clouds than sun",
-        temperature: 69,
-        temperatureDesc: "Cool",
-        comfort: 68.72,
-        humidity: 68,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 11,
-        windDirection: 283,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.92,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561410000,
-        localTime: "1406242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "More clouds than sun. Cool.",
-        skyInfo: 13,
-        skyDescription: "More clouds than sun",
-        temperature: 69,
-        temperatureDesc: "Cool",
-        comfort: 68.54,
-        humidity: 68,
-        dewPoint: 58,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 10,
-        windDirection: 283,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.7,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561413600,
-        localTime: "1506242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 68,
-        temperatureDesc: "Cool",
-        comfort: 67.82,
-        humidity: 69,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 10,
-        windDirection: 284,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.47,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561417200,
-        localTime: "1606242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 67,
-        temperatureDesc: "Cool",
-        comfort: 67.1,
-        humidity: 71,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 9,
-        windDirection: 284,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 8.23,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561420800,
-        localTime: "1706242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 66,
-        temperatureDesc: "Cool",
-        comfort: 66.2,
-        humidity: 73,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 9,
-        windDirection: 285,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.36,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561424400,
-        localTime: "1806242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 65,
-        temperatureDesc: "Cool",
-        comfort: 65.3,
-        humidity: 75,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 8,
-        windDirection: 289,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.9,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561428000,
-        localTime: "1906242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Broken clouds. Cool.",
-        skyInfo: 15,
-        skyDescription: "Broken clouds",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 64.22,
-        humidity: 78,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 7,
-        windDirection: 292,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.44,
-        icon: "6",
-        iconName: "mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561431600,
-        localTime: "2006242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Mostly cloudy. Cool.",
-        skyInfo: 16,
-        skyDescription: "Mostly cloudy",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 63.5,
-        humidity: 80,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 290,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.03,
-        icon: "17",
-        iconName: "night_mostly_cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561435200,
-        localTime: "2106242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 63.5,
-        humidity: 80,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 284,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.38,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561438800,
-        localTime: "2206242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 63.14,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 5,
-        windDirection: 281,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.57,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 2,
-        weekday: "Monday",
-        utcTime: 1561442400,
-        localTime: "2306242019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 62.96,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 5,
-        windDirection: 281,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.17,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561446000,
-        localTime: "0006252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 63.32,
-        humidity: 80,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 5,
-        windDirection: 280,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.7,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561449600,
-        localTime: "0106252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 63.32,
-        humidity: 80,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 281,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.9,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561453200,
-        localTime: "0206252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 62.96,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 283,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.48,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561456800,
-        localTime: "0306252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "N",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 63.14,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 286,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.02,
-        icon: "17",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561460400,
-        localTime: "0406252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 63.32,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 287,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.21,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561464000,
-        localTime: "0506252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 63,
-        temperatureDesc: "Cool",
-        comfort: 63.32,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 4,
-        windDirection: 278,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.54,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561467600,
-        localTime: "0606252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 63.5,
-        humidity: 81,
-        dewPoint: 57,
-        precipitationProbability: 1,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 5,
-        windDirection: 263,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.13,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561471200,
-        localTime: "0706252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 63.86,
-        humidity: 80,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 5,
-        windDirection: 252,
-        windDesc: "West",
-        windDescShort: "W",
-        visibility: 7.71,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561474800,
-        localTime: "0806252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-      {
-        daylight: "D",
-        description: "Overcast. Cool.",
-        skyInfo: 18,
-        skyDescription: "Overcast",
-        temperature: 64,
-        temperatureDesc: "Cool",
-        comfort: 64.4,
-        humidity: 78,
-        dewPoint: 57,
-        precipitationProbability: 0,
-        precipitationDesc: "",
-        rainFall: "*",
-        snowFall: "*",
-        airInfo: "*",
-        airDescription: "",
-        windSpeed: 6,
-        windDirection: 243,
-        windDesc: "Southwest",
-        windDescShort: "SW",
-        visibility: 7.24,
-        icon: "7",
-        iconName: "cloudy",
-        iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-        dayOfWeek: 3,
-        weekday: "Tuesday",
-        utcTime: 1561478400,
-        localTime: "0906252019",
-        localTimeFormat: "HHMMddyyyy",
-      },
-    ],
-  ],
+  ["weather/10001/status/display_city", "New York City, NY"],
+  ["weather/92109/status/forecast", getForecast],
+  ["weather/10001/status/forecast", getForecast],
+  ["weather/92109/status/hourly", getHourly],
   ["weather/92109/status/now", { station_id: "unknown", current_temperature: null }],
-  [
-    "weather/92109/status/observation",
-    {
-      daylight: "D",
-      description: "Mostly cloudy. Mild.",
-      skyInfo: 16,
-      skyDescription: "Mostly cloudy",
-      temperature: 67,
-      temperatureDesc: "Mild",
-      comfort: 66.9,
-      highTemperature: 69,
-      lowTemperature: 63,
-      humidity: 79,
-      dewPoint: 60,
-      precipitation1H: "*",
-      precipitation3H: "*",
-      precipitation6H: "*",
-      precipitation12H: "*",
-      precipitation24H: "*",
-      precipitationDesc: "",
-      airInfo: "*",
-      airDescription: "",
-      windSpeed: 11,
-      windDirection: 280,
-      windDesc: "West",
-      windDescShort: "W",
-      barometerPressure: 29.95,
-      barometerTrend: "",
-      visibility: 10,
-      snowCover: "*",
-      icon: "6",
-      iconName: "mostly_cloudy",
-      iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
-      ageMinutes: 24,
-      activeAlerts: false,
-      country: "United States",
-      state: "California",
-      city: "San Diego",
-      latitude: 32.7915,
-      longitude: -117.2369,
-      distance: 5.02,
-      elevation: 0,
-      utcTime: 1561405260,
-      sunrise: 1561380120.848,
-      sunset: 1561431600.849,
-    },
-  ],
+  ["weather/92109/status/observation", getObservation],
+  ["weather/10001/status/observation", getObservation],
   ["weather/92109/status/sunrise", 1561380120.848],
   ["weather/92109/status/sunset", 1561431600.849],
   ["weather/92203/status/city", "Indio"],
@@ -19166,5052 +17682,7 @@ const allMessages = [
       utcTime: 1561359600,
     },
   ],
-  [
-    "weather/10001/status/hourly",
-    [
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "13.90",
-        "temperatureDesc": "Cool",
-        "comfort": "14.17",
-        "humidity": "90",
-        "dewPoint": "12.20",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.04",
-        "windDirection": "148",
-        "windDesc": "Southeast",
-        "windDescShort": "SE",
-        "visibility": "12.49",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T02:00:00.000-04:00",
-        "localTime": "0207082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "13.20",
-        "temperatureDesc": "Cool",
-        "comfort": "13.29",
-        "humidity": "92",
-        "dewPoint": "11.90",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.40",
-        "windDirection": "141",
-        "windDesc": "Southeast",
-        "windDescShort": "SE",
-        "visibility": "12.74",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T03:00:00.000-04:00",
-        "localTime": "0307082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Quite cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "12.70",
-        "temperatureDesc": "Quite cool",
-        "comfort": "12.63",
-        "humidity": "93",
-        "dewPoint": "11.60",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.76",
-        "windDirection": "136",
-        "windDesc": "Southeast",
-        "windDescShort": "SE",
-        "visibility": "11.66",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T04:00:00.000-04:00",
-        "localTime": "0407082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Quite cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "12.60",
-        "temperatureDesc": "Quite cool",
-        "comfort": "12.43",
-        "humidity": "92",
-        "dewPoint": "11.40",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "6.12",
-        "windDirection": "130",
-        "windDesc": "Southeast",
-        "windDescShort": "SE",
-        "visibility": "11.44",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T05:00:00.000-04:00",
-        "localTime": "0507082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "13.00",
-        "temperatureDesc": "Cool",
-        "comfort": "12.82",
-        "humidity": "90",
-        "dewPoint": "11.40",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "6.48",
-        "windDirection": "123",
-        "windDesc": "Southeast",
-        "windDescShort": "SE",
-        "visibility": "11.99",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T06:00:00.000-04:00",
-        "localTime": "0607082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "14.20",
-        "temperatureDesc": "Cool",
-        "comfort": "14.21",
-        "humidity": "84",
-        "dewPoint": "11.60",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "6.48",
-        "windDirection": "113",
-        "windDesc": "Southeast",
-        "windDescShort": "SE",
-        "visibility": "12.67",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T07:00:00.000-04:00",
-        "localTime": "0707082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "16.30",
-        "temperatureDesc": "Cool",
-        "comfort": "16.30",
-        "humidity": "74",
-        "dewPoint": "11.70",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "7.20",
-        "windDirection": "101",
-        "windDesc": "East",
-        "windDescShort": "E",
-        "visibility": "11.43",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T08:00:00.000-04:00",
-        "localTime": "0807082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "19.10",
-        "temperatureDesc": "Cool",
-        "comfort": "19.10",
-        "humidity": "64",
-        "dewPoint": "12.00",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.92",
-        "windDirection": "92",
-        "windDesc": "East",
-        "windDescShort": "E",
-        "visibility": "12.50",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T09:00:00.000-04:00",
-        "localTime": "0907082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "21.30",
-        "temperatureDesc": "Cool",
-        "comfort": "24.73",
-        "humidity": "56",
-        "dewPoint": "12.20",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "8.28",
-        "windDirection": "80",
-        "windDesc": "East",
-        "windDescShort": "E",
-        "visibility": "13.41",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T10:00:00.000-04:00",
-        "localTime": "1007082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "23.00",
-        "temperatureDesc": "Mild",
-        "comfort": "25.08",
-        "humidity": "51",
-        "dewPoint": "12.30",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "9.00",
-        "windDirection": "71",
-        "windDesc": "East",
-        "windDescShort": "E",
-        "visibility": "13.09",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T11:00:00.000-04:00",
-        "localTime": "1107082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "24.30",
-        "temperatureDesc": "Mild",
-        "comfort": "25.51",
-        "humidity": "48",
-        "dewPoint": "12.50",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "9.00",
-        "windDirection": "60",
-        "windDesc": "Northeast",
-        "windDescShort": "NE",
-        "visibility": "15.23",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T12:00:00.000-04:00",
-        "localTime": "1207082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "25.30",
-        "temperatureDesc": "Mild",
-        "comfort": "25.98",
-        "humidity": "46",
-        "dewPoint": "12.90",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.00",
-        "windDirection": "37",
-        "windDesc": "Northeast",
-        "windDescShort": "NE",
-        "visibility": "15.56",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T13:00:00.000-04:00",
-        "localTime": "1307082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "26.00",
-        "temperatureDesc": "Mild",
-        "comfort": "26.38",
-        "humidity": "45",
-        "dewPoint": "13.30",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "8.64",
-        "windDirection": "6",
-        "windDesc": "North",
-        "windDescShort": "N",
-        "visibility": "15.46",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T14:00:00.000-04:00",
-        "localTime": "1407082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "26.40",
-        "temperatureDesc": "Mild",
-        "comfort": "26.62",
-        "humidity": "45",
-        "dewPoint": "13.50",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.72",
-        "windDirection": "341",
-        "windDesc": "North",
-        "windDescShort": "N",
-        "visibility": "18.55",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T15:00:00.000-04:00",
-        "localTime": "1507082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "26.50",
-        "temperatureDesc": "Mild",
-        "comfort": "26.69",
-        "humidity": "44",
-        "dewPoint": "13.40",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.44",
-        "windDirection": "324",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "19.94",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T16:00:00.000-04:00",
-        "localTime": "1607082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "26.00",
-        "temperatureDesc": "Mild",
-        "comfort": "26.38",
-        "humidity": "46",
-        "dewPoint": "13.40",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "21",
-        "airDescription": "Very nice",
-        "windSpeed": "10.80",
-        "windDirection": "314",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "21.30",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T17:00:00.000-04:00",
-        "localTime": "1707082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "25.30",
-        "temperatureDesc": "Mild",
-        "comfort": "26.03",
-        "humidity": "48",
-        "dewPoint": "13.60",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.80",
-        "windDirection": "306",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "15.34",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T18:00:00.000-04:00",
-        "localTime": "1807082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "24.30",
-        "temperatureDesc": "Mild",
-        "comfort": "25.52",
-        "humidity": "52",
-        "dewPoint": "13.90",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "10.80",
-        "windDirection": "298",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "14.68",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T19:00:00.000-04:00",
-        "localTime": "1907082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "22.30",
-        "temperatureDesc": "Mild",
-        "comfort": "24.53",
-        "humidity": "60",
-        "dewPoint": "14.20",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "10.44",
-        "windDirection": "292",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "15.10",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T20:00:00.000-04:00",
-        "localTime": "2007082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "20.80",
-        "temperatureDesc": "Cool",
-        "comfort": "20.80",
-        "humidity": "65",
-        "dewPoint": "14.00",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "9.00",
-        "windDirection": "284",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "13.98",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T21:00:00.000-04:00",
-        "localTime": "2107082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "18.90",
-        "temperatureDesc": "Cool",
-        "comfort": "18.90",
-        "humidity": "69",
-        "dewPoint": "13.00",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.56",
-        "windDirection": "272",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "11.34",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T22:00:00.000-04:00",
-        "localTime": "2207082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "16.80",
-        "temperatureDesc": "Cool",
-        "comfort": "16.80",
-        "humidity": "74",
-        "dewPoint": "12.20",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.48",
-        "windDirection": "260",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "12.69",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-08T23:00:00.000-04:00",
-        "localTime": "2307082019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "15.10",
-        "temperatureDesc": "Cool",
-        "comfort": "15.31",
-        "humidity": "80",
-        "dewPoint": "11.70",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.12",
-        "windDirection": "253",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "11.73",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T00:00:00.000-04:00",
-        "localTime": "0007092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "13.80",
-        "temperatureDesc": "Cool",
-        "comfort": "13.89",
-        "humidity": "85",
-        "dewPoint": "11.30",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.76",
-        "windDirection": "245",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "11.90",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T01:00:00.000-04:00",
-        "localTime": "0107092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "12.80",
-        "temperatureDesc": "Cool",
-        "comfort": "12.83",
-        "humidity": "88",
-        "dewPoint": "10.90",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.40",
-        "windDirection": "238",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "12.66",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T02:00:00.000-04:00",
-        "localTime": "0207092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Quite cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "12.00",
-        "temperatureDesc": "Quite cool",
-        "comfort": "11.92",
-        "humidity": "91",
-        "dewPoint": "10.50",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.40",
-        "windDirection": "237",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "11.69",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T03:00:00.000-04:00",
-        "localTime": "0307092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Quite cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "11.50",
-        "temperatureDesc": "Quite cool",
-        "comfort": "11.26",
-        "humidity": "92",
-        "dewPoint": "10.20",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.76",
-        "windDirection": "243",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "12.35",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T04:00:00.000-04:00",
-        "localTime": "0407092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Quite cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "11.20",
-        "temperatureDesc": "Quite cool",
-        "comfort": "10.91",
-        "humidity": "93",
-        "dewPoint": "10.10",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.76",
-        "windDirection": "252",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "11.38",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T05:00:00.000-04:00",
-        "localTime": "0507092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Quite cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "11.90",
-        "temperatureDesc": "Quite cool",
-        "comfort": "11.80",
-        "humidity": "91",
-        "dewPoint": "10.50",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "5.40",
-        "windDirection": "266",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "12.26",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T06:00:00.000-04:00",
-        "localTime": "0607092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "14.00",
-        "temperatureDesc": "Cool",
-        "comfort": "14.00",
-        "humidity": "86",
-        "dewPoint": "11.60",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "4.68",
-        "windDirection": "288",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "12.81",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T07:00:00.000-04:00",
-        "localTime": "0707092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "16.40",
-        "temperatureDesc": "Cool",
-        "comfort": "16.40",
-        "humidity": "77",
-        "dewPoint": "12.40",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "4.68",
-        "windDirection": "306",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "12.85",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T08:00:00.000-04:00",
-        "localTime": "0807092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "19.40",
-        "temperatureDesc": "Cool",
-        "comfort": "19.40",
-        "humidity": "65",
-        "dewPoint": "12.70",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "6.48",
-        "windDirection": "314",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "12.65",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T09:00:00.000-04:00",
-        "localTime": "0907092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Cool.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "21.70",
-        "temperatureDesc": "Cool",
-        "comfort": "24.68",
-        "humidity": "56",
-        "dewPoint": "12.60",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "7.92",
-        "windDirection": "316",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "11.61",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T10:00:00.000-04:00",
-        "localTime": "1007092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "23.70",
-        "temperatureDesc": "Mild",
-        "comfort": "25.28",
-        "humidity": "49",
-        "dewPoint": "12.40",
-        "precipitationProbability": "0",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "9.00",
-        "windDirection": "314",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "13.28",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T11:00:00.000-04:00",
-        "localTime": "1107092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "25.20",
-        "temperatureDesc": "Mild",
-        "comfort": "25.88",
-        "humidity": "44",
-        "dewPoint": "12.20",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "9.72",
-        "windDirection": "307",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "14.17",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T12:00:00.000-04:00",
-        "localTime": "1207092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Mild.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "26.30",
-        "temperatureDesc": "Mild",
-        "comfort": "26.43",
-        "humidity": "41",
-        "dewPoint": "12.00",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "10.08",
-        "windDirection": "298",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "14.87",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T13:00:00.000-04:00",
-        "localTime": "1307092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Warm.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "27.00",
-        "temperatureDesc": "Warm",
-        "comfort": "26.81",
-        "humidity": "39",
-        "dewPoint": "11.90",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.44",
-        "windDirection": "292",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "15.47",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T14:00:00.000-04:00",
-        "localTime": "1407092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Warm.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "27.60",
-        "temperatureDesc": "Warm",
-        "comfort": "27.21",
-        "humidity": "38",
-        "dewPoint": "12.00",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.80",
-        "windDirection": "290",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "16.01",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T15:00:00.000-04:00",
-        "localTime": "1507092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Warm.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "27.70",
-        "temperatureDesc": "Warm",
-        "comfort": "27.29",
-        "humidity": "39",
-        "dewPoint": "12.40",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.80",
-        "windDirection": "290",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "15.31",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T16:00:00.000-04:00",
-        "localTime": "1607092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Warm.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "27.30",
-        "temperatureDesc": "Warm",
-        "comfort": "27.09",
-        "humidity": "41",
-        "dewPoint": "12.80",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.44",
-        "windDirection": "290",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "14.77",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T17:00:00.000-04:00",
-        "localTime": "1707092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "26.50",
-        "temperatureDesc": "Mild",
-        "comfort": "26.69",
-        "humidity": "45",
-        "dewPoint": "13.50",
-        "precipitationProbability": "1",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "9.72",
-        "windDirection": "293",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "15.78",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T18:00:00.000-04:00",
-        "localTime": "1807092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "25.30",
-        "temperatureDesc": "Mild",
-        "comfort": "26.07",
-        "humidity": "51",
-        "dewPoint": "14.40",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "7.92",
-        "windDirection": "298",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "13.11",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T19:00:00.000-04:00",
-        "localTime": "1907092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Sunny. Mild.",
-        "skyInfo": "1",
-        "skyDescription": "Sunny",
-        "temperature": "23.40",
-        "temperatureDesc": "Mild",
-        "comfort": "24.97",
-        "humidity": "60",
-        "dewPoint": "15.10",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "20",
-        "airDescription": "Beautiful",
-        "windSpeed": "6.84",
-        "windDirection": "298",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "13.40",
-        "icon": "1",
-        "iconName": "sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/1.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T20:00:00.000-04:00",
-        "localTime": "2007092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "22.00",
-        "temperatureDesc": "Cool",
-        "comfort": "23.97",
-        "humidity": "67",
-        "dewPoint": "15.50",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.84",
-        "windDirection": "257",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "12.81",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T21:00:00.000-04:00",
-        "localTime": "2107092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "20.50",
-        "temperatureDesc": "Cool",
-        "comfort": "20.50",
-        "humidity": "71",
-        "dewPoint": "15.10",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.84",
-        "windDirection": "195",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "11.79",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T22:00:00.000-04:00",
-        "localTime": "2207092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "18.40",
-        "temperatureDesc": "Cool",
-        "comfort": "18.40",
-        "humidity": "78",
-        "dewPoint": "14.50",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.84",
-        "windDirection": "180",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "11.98",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "3",
-        "weekday": "Tuesday",
-        "utcTime": "2019-07-09T23:00:00.000-04:00",
-        "localTime": "2307092019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "16.50",
-        "temperatureDesc": "Cool",
-        "comfort": "16.50",
-        "humidity": "85",
-        "dewPoint": "13.90",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "7.56",
-        "windDirection": "177",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "12.44",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T00:00:00.000-04:00",
-        "localTime": "0007102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "15.40",
-        "temperatureDesc": "Cool",
-        "comfort": "15.31",
-        "humidity": "88",
-        "dewPoint": "13.40",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "8.28",
-        "windDirection": "179",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "11.84",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T01:00:00.000-04:00",
-        "localTime": "0107102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "15.70",
-        "temperatureDesc": "Cool",
-        "comfort": "15.70",
-        "humidity": "84",
-        "dewPoint": "13.00",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "8.64",
-        "windDirection": "179",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "12.46",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T02:00:00.000-04:00",
-        "localTime": "0207102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "16.30",
-        "temperatureDesc": "Cool",
-        "comfort": "16.30",
-        "humidity": "79",
-        "dewPoint": "12.70",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "8.64",
-        "windDirection": "178",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "12.24",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T03:00:00.000-04:00",
-        "localTime": "0307102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Clear. Cool.",
-        "skyInfo": "2",
-        "skyDescription": "Clear",
-        "temperature": "16.20",
-        "temperatureDesc": "Cool",
-        "comfort": "16.20",
-        "humidity": "78",
-        "dewPoint": "12.40",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "8.28",
-        "windDirection": "176",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "11.93",
-        "icon": "13",
-        "iconName": "night_clear",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/16.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T04:00:00.000-04:00",
-        "localTime": "0407102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "15.30",
-        "temperatureDesc": "Cool",
-        "comfort": "15.24",
-        "humidity": "82",
-        "dewPoint": "12.30",
-        "precipitationProbability": "2",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "7.92",
-        "windDirection": "175",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "12.52",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T05:00:00.000-04:00",
-        "localTime": "0507102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "14.80",
-        "temperatureDesc": "Cool",
-        "comfort": "14.66",
-        "humidity": "88",
-        "dewPoint": "12.80",
-        "precipitationProbability": "3",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "7.92",
-        "windDirection": "175",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "12.24",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T06:00:00.000-04:00",
-        "localTime": "0607102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "16.60",
-        "temperatureDesc": "Cool",
-        "comfort": "16.60",
-        "humidity": "85",
-        "dewPoint": "14.10",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "7.56",
-        "windDirection": "178",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "11.31",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T07:00:00.000-04:00",
-        "localTime": "0707102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "18.70",
-        "temperatureDesc": "Cool",
-        "comfort": "18.70",
-        "humidity": "81",
-        "dewPoint": "15.40",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "32",
-        "airDescription": "Damp",
-        "windSpeed": "7.20",
-        "windDirection": "182",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "11.92",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T08:00:00.000-04:00",
-        "localTime": "0807102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Cool.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "21.60",
-        "temperatureDesc": "Cool",
-        "comfort": "23.40",
-        "humidity": "71",
-        "dewPoint": "16.00",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.20",
-        "windDirection": "198",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "12.27",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T09:00:00.000-04:00",
-        "localTime": "0907102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Mild.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "24.10",
-        "temperatureDesc": "Mild",
-        "comfort": "25.28",
-        "humidity": "63",
-        "dewPoint": "16.50",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.84",
-        "windDirection": "227",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "11.84",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T10:00:00.000-04:00",
-        "localTime": "1007102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Mild.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "26.40",
-        "temperatureDesc": "Mild",
-        "comfort": "27.07",
-        "humidity": "55",
-        "dewPoint": "16.70",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.20",
-        "windDirection": "248",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "11.89",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T11:00:00.000-04:00",
-        "localTime": "1107102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Warm.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "28.00",
-        "temperatureDesc": "Warm",
-        "comfort": "28.45",
-        "humidity": "50",
-        "dewPoint": "16.60",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.56",
-        "windDirection": "258",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "14.43",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T12:00:00.000-04:00",
-        "localTime": "1207102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Warm.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "28.90",
-        "temperatureDesc": "Warm",
-        "comfort": "29.19",
-        "humidity": "47",
-        "dewPoint": "16.50",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "8.28",
-        "windDirection": "264",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "13.32",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T13:00:00.000-04:00",
-        "localTime": "1307102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Mostly sunny. Warm.",
-        "skyInfo": "3",
-        "skyDescription": "Mostly sunny",
-        "temperature": "29.40",
-        "temperatureDesc": "Warm",
-        "comfort": "29.56",
-        "humidity": "45",
-        "dewPoint": "16.40",
-        "precipitationProbability": "3",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "8.64",
-        "windDirection": "268",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "15.35",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T14:00:00.000-04:00",
-        "localTime": "1407102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Passing clouds. Warm.",
-        "skyInfo": "7",
-        "skyDescription": "Passing clouds",
-        "temperature": "29.80",
-        "temperatureDesc": "Warm",
-        "comfort": "29.93",
-        "humidity": "45",
-        "dewPoint": "16.50",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.00",
-        "windDirection": "268",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T15:00:00.000-04:00",
-        "localTime": "1507102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Passing clouds. Warm.",
-        "skyInfo": "7",
-        "skyDescription": "Passing clouds",
-        "temperature": "30.00",
-        "temperatureDesc": "Warm",
-        "comfort": "30.18",
-        "humidity": "44",
-        "dewPoint": "16.60",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.00",
-        "windDirection": "258",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T16:00:00.000-04:00",
-        "localTime": "1607102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "More sun than clouds. Warm.",
-        "skyInfo": "8",
-        "skyDescription": "More sun than clouds",
-        "temperature": "29.60",
-        "temperatureDesc": "Warm",
-        "comfort": "29.93",
-        "humidity": "46",
-        "dewPoint": "16.90",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "21",
-        "airDescription": "Very nice",
-        "windSpeed": "8.64",
-        "windDirection": "241",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T17:00:00.000-04:00",
-        "localTime": "1707102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "More sun than clouds. Warm.",
-        "skyInfo": "8",
-        "skyDescription": "More sun than clouds",
-        "temperature": "29.00",
-        "temperatureDesc": "Warm",
-        "comfort": "29.54",
-        "humidity": "50",
-        "dewPoint": "17.40",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "8.64",
-        "windDirection": "223",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T18:00:00.000-04:00",
-        "localTime": "1807102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "More sun than clouds. Warm.",
-        "skyInfo": "8",
-        "skyDescription": "More sun than clouds",
-        "temperature": "27.80",
-        "temperatureDesc": "Warm",
-        "comfort": "28.67",
-        "humidity": "56",
-        "dewPoint": "18.10",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "8.64",
-        "windDirection": "206",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T19:00:00.000-04:00",
-        "localTime": "1907102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "More sun than clouds. Mild.",
-        "skyInfo": "8",
-        "skyDescription": "More sun than clouds",
-        "temperature": "25.70",
-        "temperatureDesc": "Mild",
-        "comfort": "26.73",
-        "humidity": "65",
-        "dewPoint": "18.50",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "21",
-        "airDescription": "Very nice",
-        "windSpeed": "8.64",
-        "windDirection": "195",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T20:00:00.000-04:00",
-        "localTime": "2007102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Mild.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "24.80",
-        "temperatureDesc": "Mild",
-        "comfort": "25.78",
-        "humidity": "69",
-        "dewPoint": "18.60",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.36",
-        "windDirection": "190",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T21:00:00.000-04:00",
-        "localTime": "2107102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Mild.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "23.40",
-        "temperatureDesc": "Mild",
-        "comfort": "24.27",
-        "humidity": "72",
-        "dewPoint": "18.10",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.44",
-        "windDirection": "190",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T22:00:00.000-04:00",
-        "localTime": "2207102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "21.40",
-        "temperatureDesc": "Cool",
-        "comfort": "21.95",
-        "humidity": "80",
-        "dewPoint": "17.70",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "11.16",
-        "windDirection": "188",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "4",
-        "weekday": "Wednesday",
-        "utcTime": "2019-07-10T23:00:00.000-04:00",
-        "localTime": "2307102019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "21.40",
-        "temperatureDesc": "Cool",
-        "comfort": "21.95",
-        "humidity": "80",
-        "dewPoint": "17.70",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "11.88",
-        "windDirection": "187",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T00:00:00.000-04:00",
-        "localTime": "0007112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "21.70",
-        "temperatureDesc": "Cool",
-        "comfort": "22.14",
-        "humidity": "79",
-        "dewPoint": "17.90",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "12.60",
-        "windDirection": "185",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T01:00:00.000-04:00",
-        "localTime": "0107112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "20.60",
-        "temperatureDesc": "Cool",
-        "comfort": "20.60",
-        "humidity": "85",
-        "dewPoint": "17.90",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "12.60",
-        "windDirection": "184",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T02:00:00.000-04:00",
-        "localTime": "0207112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.50",
-        "temperatureDesc": "Cool",
-        "comfort": "20.50",
-        "humidity": "86",
-        "dewPoint": "18.00",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "12.60",
-        "windDirection": "184",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "17",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T03:00:00.000-04:00",
-        "localTime": "0307112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.60",
-        "temperatureDesc": "Cool",
-        "comfort": "20.60",
-        "humidity": "86",
-        "dewPoint": "18.10",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "12.60",
-        "windDirection": "186",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "17",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T04:00:00.000-04:00",
-        "localTime": "0407112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.20",
-        "temperatureDesc": "Cool",
-        "comfort": "20.20",
-        "humidity": "88",
-        "dewPoint": "18.10",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "12.60",
-        "windDirection": "187",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T05:00:00.000-04:00",
-        "localTime": "0507112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a few showers. Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.30",
-        "temperatureDesc": "Cool",
-        "comfort": "20.30",
-        "humidity": "87",
-        "dewPoint": "18.10",
-        "precipitationProbability": "33",
-        "precipitationDesc": "a few showers",
-        "rainFall": "0.02",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "12.96",
-        "windDirection": "186",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T06:00:00.000-04:00",
-        "localTime": "0607112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a few showers. Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.40",
-        "temperatureDesc": "Cool",
-        "comfort": "20.40",
-        "humidity": "87",
-        "dewPoint": "18.10",
-        "precipitationProbability": "45",
-        "precipitationDesc": "a few showers",
-        "rainFall": "0.02",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "13.32",
-        "windDirection": "183",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T07:00:00.000-04:00",
-        "localTime": "0707112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a few showers. Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.80",
-        "temperatureDesc": "Cool",
-        "comfort": "20.80",
-        "humidity": "84",
-        "dewPoint": "18.00",
-        "precipitationProbability": "57",
-        "precipitationDesc": "a few showers",
-        "rainFall": "0.03",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "13.68",
-        "windDirection": "179",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T08:00:00.000-04:00",
-        "localTime": "0807112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "21.90",
-        "temperatureDesc": "Cool",
-        "comfort": "22.43",
-        "humidity": "79",
-        "dewPoint": "18.10",
-        "precipitationProbability": "51",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.12",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "14.40",
-        "windDirection": "178",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T09:00:00.000-04:00",
-        "localTime": "0907112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Cloudy. Mild.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "23.90",
-        "temperatureDesc": "Mild",
-        "comfort": "24.83",
-        "humidity": "71",
-        "dewPoint": "18.30",
-        "precipitationProbability": "45",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.10",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "16.20",
-        "windDirection": "177",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T10:00:00.000-04:00",
-        "localTime": "1007112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light rain. Cloudy. Mild.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "25.30",
-        "temperatureDesc": "Mild",
-        "comfort": "26.31",
-        "humidity": "66",
-        "dewPoint": "18.40",
-        "precipitationProbability": "40",
-        "precipitationDesc": "Light rain",
-        "rainFall": "0.09",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "17.28",
-        "windDirection": "179",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "18",
-        "iconName": "sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/27.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T11:00:00.000-04:00",
-        "localTime": "1107112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Rain. Cloudy. Mild.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "26.40",
-        "temperatureDesc": "Mild",
-        "comfort": "27.46",
-        "humidity": "62",
-        "dewPoint": "18.50",
-        "precipitationProbability": "45",
-        "precipitationDesc": "Rain",
-        "rainFall": "0.18",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "18.36",
-        "windDirection": "181",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "19",
-        "iconName": "rain",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/27.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T12:00:00.000-04:00",
-        "localTime": "1207112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Rain. Cloudy. Warm.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "27.30",
-        "temperatureDesc": "Warm",
-        "comfort": "28.31",
-        "humidity": "59",
-        "dewPoint": "18.50",
-        "precipitationProbability": "50",
-        "precipitationDesc": "Rain",
-        "rainFall": "0.20",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "19.08",
-        "windDirection": "185",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "19",
-        "iconName": "rain",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/27.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T13:00:00.000-04:00",
-        "localTime": "1307112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Rain. Cloudy. Warm.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "28.00",
-        "temperatureDesc": "Warm",
-        "comfort": "29.02",
-        "humidity": "56",
-        "dewPoint": "18.50",
-        "precipitationProbability": "55",
-        "precipitationDesc": "Rain",
-        "rainFall": "0.22",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "19.08",
-        "windDirection": "189",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "19",
-        "iconName": "rain",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/27.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T14:00:00.000-04:00",
-        "localTime": "1407112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Cloudy. Warm.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "28.30",
-        "temperatureDesc": "Warm",
-        "comfort": "29.32",
-        "humidity": "55",
-        "dewPoint": "18.40",
-        "precipitationProbability": "48",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.13",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "19.08",
-        "windDirection": "190",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T15:00:00.000-04:00",
-        "localTime": "1507112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Overcast. Warm.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "28.50",
-        "temperatureDesc": "Warm",
-        "comfort": "29.59",
-        "humidity": "55",
-        "dewPoint": "18.70",
-        "precipitationProbability": "41",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.11",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "19.08",
-        "windDirection": "190",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T16:00:00.000-04:00",
-        "localTime": "1607112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Overcast. Warm.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "28.10",
-        "temperatureDesc": "Warm",
-        "comfort": "29.37",
-        "humidity": "58",
-        "dewPoint": "19.10",
-        "precipitationProbability": "34",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.09",
-        "snowFall": "*",
-        "airInfo": "34",
-        "airDescription": "Muggy",
-        "windSpeed": "18.72",
-        "windDirection": "191",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T17:00:00.000-04:00",
-        "localTime": "1707112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Overcast. Warm.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "27.50",
-        "temperatureDesc": "Warm",
-        "comfort": "28.92",
-        "humidity": "62",
-        "dewPoint": "19.60",
-        "precipitationProbability": "41",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.10",
-        "snowFall": "*",
-        "airInfo": "34",
-        "airDescription": "Muggy",
-        "windSpeed": "17.64",
-        "windDirection": "202",
-        "windDesc": "South",
-        "windDescShort": "S",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T18:00:00.000-04:00",
-        "localTime": "1807112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Overcast. Mild.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "26.50",
-        "temperatureDesc": "Mild",
-        "comfort": "27.87",
-        "humidity": "68",
-        "dewPoint": "20.10",
-        "precipitationProbability": "48",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.12",
-        "snowFall": "*",
-        "airInfo": "34",
-        "airDescription": "Muggy",
-        "windSpeed": "16.92",
-        "windDirection": "219",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T19:00:00.000-04:00",
-        "localTime": "1907112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Light showers. Overcast. Mild.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "25.20",
-        "temperatureDesc": "Mild",
-        "comfort": "26.23",
-        "humidity": "76",
-        "dewPoint": "20.60",
-        "precipitationProbability": "55",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.14",
-        "snowFall": "*",
-        "airInfo": "34",
-        "airDescription": "Muggy",
-        "windSpeed": "15.12",
-        "windDirection": "232",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "33",
-        "iconName": "showers",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/10.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T20:00:00.000-04:00",
-        "localTime": "2007112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "a few showers. Overcast. Mild.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "24.20",
-        "temperatureDesc": "Mild",
-        "comfort": "24.53",
-        "humidity": "82",
-        "dewPoint": "20.90",
-        "precipitationProbability": "50",
-        "precipitationDesc": "a few showers",
-        "rainFall": "0.01",
-        "snowFall": "*",
-        "airInfo": "34",
-        "airDescription": "Muggy",
-        "windSpeed": "12.96",
-        "windDirection": "237",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T21:00:00.000-04:00",
-        "localTime": "2107112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "a few showers. Overcast. Mild.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "22.50",
-        "temperatureDesc": "Mild",
-        "comfort": "21.25",
-        "humidity": "89",
-        "dewPoint": "20.60",
-        "precipitationProbability": "46",
-        "precipitationDesc": "a few showers",
-        "rainFall": "0.01",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "10.80",
-        "windDirection": "242",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T22:00:00.000-04:00",
-        "localTime": "2207112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "a few showers. Overcast. Cool.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "21.00",
-        "temperatureDesc": "Cool",
-        "comfort": "21.00",
-        "humidity": "96",
-        "dewPoint": "20.30",
-        "precipitationProbability": "41",
-        "precipitationDesc": "a few showers",
-        "rainFall": "0.01",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "9.72",
-        "windDirection": "244",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "5",
-        "weekday": "Thursday",
-        "utcTime": "2019-07-11T23:00:00.000-04:00",
-        "localTime": "2307112019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Light showers. Overcast. Cool.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "20.40",
-        "temperatureDesc": "Cool",
-        "comfort": "20.40",
-        "humidity": "98",
-        "dewPoint": "20.00",
-        "precipitationProbability": "44",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.11",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "9.72",
-        "windDirection": "245",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T00:00:00.000-04:00",
-        "localTime": "0007122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Light showers. Overcast. Cool.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "20.30",
-        "temperatureDesc": "Cool",
-        "comfort": "20.30",
-        "humidity": "96",
-        "dewPoint": "19.70",
-        "precipitationProbability": "47",
-        "precipitationDesc": "Light showers",
-        "rainFall": "0.11",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "10.08",
-        "windDirection": "244",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T01:00:00.000-04:00",
-        "localTime": "0107122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Light rain. Overcast. Cool.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "19.80",
-        "temperatureDesc": "Cool",
-        "comfort": "19.80",
-        "humidity": "97",
-        "dewPoint": "19.30",
-        "precipitationProbability": "50",
-        "precipitationDesc": "Light rain",
-        "rainFall": "0.12",
-        "snowFall": "*",
-        "airInfo": "33",
-        "airDescription": "Humid",
-        "windSpeed": "10.44",
-        "windDirection": "244",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T02:00:00.000-04:00",
-        "localTime": "0207122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Drizzle. Overcast. Cool.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "19.30",
-        "temperatureDesc": "Cool",
-        "comfort": "19.30",
-        "humidity": "98",
-        "dewPoint": "18.90",
-        "precipitationProbability": "39",
-        "precipitationDesc": "Drizzle",
-        "rainFall": "0.02",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "10.08",
-        "windDirection": "243",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T03:00:00.000-04:00",
-        "localTime": "0307122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Drizzle. Overcast. Cool.",
-        "skyInfo": "18",
-        "skyDescription": "Overcast",
-        "temperature": "19.00",
-        "temperatureDesc": "Cool",
-        "comfort": "19.00",
-        "humidity": "98",
-        "dewPoint": "18.60",
-        "precipitationProbability": "28",
-        "precipitationDesc": "Drizzle",
-        "rainFall": "0.01",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "9.72",
-        "windDirection": "239",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "34",
-        "iconName": "night_sprinkles",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/22.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T04:00:00.000-04:00",
-        "localTime": "0407122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "18.80",
-        "temperatureDesc": "Cool",
-        "comfort": "18.80",
-        "humidity": "97",
-        "dewPoint": "18.30",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "9.72",
-        "windDirection": "240",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T05:00:00.000-04:00",
-        "localTime": "0507122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "18.70",
-        "temperatureDesc": "Cool",
-        "comfort": "18.70",
-        "humidity": "98",
-        "dewPoint": "18.30",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "10.44",
-        "windDirection": "243",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T06:00:00.000-04:00",
-        "localTime": "0607122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "19.20",
-        "temperatureDesc": "Cool",
-        "comfort": "19.20",
-        "humidity": "98",
-        "dewPoint": "18.90",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "11.16",
-        "windDirection": "246",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T07:00:00.000-04:00",
-        "localTime": "0707122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.00",
-        "temperatureDesc": "Cool",
-        "comfort": "20.00",
-        "humidity": "95",
-        "dewPoint": "19.20",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "11.88",
-        "windDirection": "250",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T08:00:00.000-04:00",
-        "localTime": "0807122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "20.70",
-        "temperatureDesc": "Cool",
-        "comfort": "20.70",
-        "humidity": "93",
-        "dewPoint": "19.50",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "12.60",
-        "windDirection": "258",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T09:00:00.000-04:00",
-        "localTime": "0907122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "21.10",
-        "temperatureDesc": "Cool",
-        "comfort": "21.10",
-        "humidity": "92",
-        "dewPoint": "19.70",
-        "precipitationProbability": "8",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "14.04",
-        "windDirection": "269",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T10:00:00.000-04:00",
-        "localTime": "1007122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "21.40",
-        "temperatureDesc": "Cool",
-        "comfort": "19.69",
-        "humidity": "91",
-        "dewPoint": "19.80",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "15.48",
-        "windDirection": "279",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T11:00:00.000-04:00",
-        "localTime": "1107122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "21.60",
-        "temperatureDesc": "Cool",
-        "comfort": "20.14",
-        "humidity": "90",
-        "dewPoint": "19.80",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "16.92",
-        "windDirection": "285",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T12:00:00.000-04:00",
-        "localTime": "1207122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Cloudy. Cool.",
-        "skyInfo": "17",
-        "skyDescription": "Cloudy",
-        "temperature": "21.80",
-        "temperatureDesc": "Cool",
-        "comfort": "20.77",
-        "humidity": "88",
-        "dewPoint": "19.80",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "18.72",
-        "windDirection": "291",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "7",
-        "iconName": "cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/17.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T13:00:00.000-04:00",
-        "localTime": "1307122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.00",
-        "temperatureDesc": "Cool",
-        "comfort": "20.99",
-        "humidity": "87",
-        "dewPoint": "19.80",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "19.80",
-        "windDirection": "296",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T14:00:00.000-04:00",
-        "localTime": "1407122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.30",
-        "temperatureDesc": "Mild",
-        "comfort": "21.51",
-        "humidity": "86",
-        "dewPoint": "19.90",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "20.52",
-        "windDirection": "300",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T15:00:00.000-04:00",
-        "localTime": "1507122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.60",
-        "temperatureDesc": "Mild",
-        "comfort": "22.17",
-        "humidity": "85",
-        "dewPoint": "19.90",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "20.16",
-        "windDirection": "303",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T16:00:00.000-04:00",
-        "localTime": "1607122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.80",
-        "temperatureDesc": "Mild",
-        "comfort": "22.68",
-        "humidity": "83",
-        "dewPoint": "19.80",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "19.44",
-        "windDirection": "304",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T17:00:00.000-04:00",
-        "localTime": "1707122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.40",
-        "temperatureDesc": "Mild",
-        "comfort": "21.95",
-        "humidity": "85",
-        "dewPoint": "19.70",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "18.00",
-        "windDirection": "303",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T18:00:00.000-04:00",
-        "localTime": "1807122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "21.60",
-        "temperatureDesc": "Cool",
-        "comfort": "20.77",
-        "humidity": "87",
-        "dewPoint": "19.40",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "16.20",
-        "windDirection": "301",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T19:00:00.000-04:00",
-        "localTime": "1907122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "20.60",
-        "temperatureDesc": "Cool",
-        "comfort": "20.60",
-        "humidity": "91",
-        "dewPoint": "19.00",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "14.04",
-        "windDirection": "299",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T20:00:00.000-04:00",
-        "localTime": "2007122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "20.30",
-        "temperatureDesc": "Cool",
-        "comfort": "20.30",
-        "humidity": "90",
-        "dewPoint": "18.60",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "11.52",
-        "windDirection": "295",
-        "windDesc": "Northwest",
-        "windDescShort": "NW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T21:00:00.000-04:00",
-        "localTime": "2107122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.60",
-        "temperatureDesc": "Cool",
-        "comfort": "19.60",
-        "humidity": "92",
-        "dewPoint": "18.30",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.08",
-        "windDirection": "289",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T22:00:00.000-04:00",
-        "localTime": "2207122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "18.70",
-        "temperatureDesc": "Cool",
-        "comfort": "18.70",
-        "humidity": "95",
-        "dewPoint": "17.90",
-        "precipitationProbability": "3",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.00",
-        "windDirection": "282",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "6",
-        "weekday": "Friday",
-        "utcTime": "2019-07-12T23:00:00.000-04:00",
-        "localTime": "2307122019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "18.30",
-        "temperatureDesc": "Cool",
-        "comfort": "18.30",
-        "humidity": "96",
-        "dewPoint": "17.70",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.00",
-        "windDirection": "277",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T00:00:00.000-04:00",
-        "localTime": "0007132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "18.30",
-        "temperatureDesc": "Cool",
-        "comfort": "18.30",
-        "humidity": "96",
-        "dewPoint": "17.60",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.72",
-        "windDirection": "275",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T01:00:00.000-04:00",
-        "localTime": "0107132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "18.20",
-        "temperatureDesc": "Cool",
-        "comfort": "18.20",
-        "humidity": "96",
-        "dewPoint": "17.60",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "9.72",
-        "windDirection": "275",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T02:00:00.000-04:00",
-        "localTime": "0207132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "18.00",
-        "temperatureDesc": "Cool",
-        "comfort": "18.00",
-        "humidity": "98",
-        "dewPoint": "17.60",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "9.36",
-        "windDirection": "276",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T03:00:00.000-04:00",
-        "localTime": "0307132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "17.80",
-        "temperatureDesc": "Cool",
-        "comfort": "17.80",
-        "humidity": "98",
-        "dewPoint": "17.40",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "9.00",
-        "windDirection": "277",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T04:00:00.000-04:00",
-        "localTime": "0407132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "17.40",
-        "temperatureDesc": "Cool",
-        "comfort": "17.40",
-        "humidity": "98",
-        "dewPoint": "17.00",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "9.00",
-        "windDirection": "279",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T05:00:00.000-04:00",
-        "localTime": "0507132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "17.00",
-        "temperatureDesc": "Cool",
-        "comfort": "17.00",
-        "humidity": "97",
-        "dewPoint": "16.50",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "8.64",
-        "windDirection": "279",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T06:00:00.000-04:00",
-        "localTime": "0607132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "15.70",
-        "temperatureDesc": "Cool",
-        "comfort": "15.70",
-        "humidity": "96",
-        "dewPoint": "15.00",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "9.00",
-        "windDirection": "280",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T07:00:00.000-04:00",
-        "localTime": "0707132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "17.40",
-        "temperatureDesc": "Cool",
-        "comfort": "17.40",
-        "humidity": "83",
-        "dewPoint": "14.50",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "22",
-        "airDescription": "Nice",
-        "windSpeed": "9.36",
-        "windDirection": "280",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T08:00:00.000-04:00",
-        "localTime": "0807132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "19.10",
-        "temperatureDesc": "Cool",
-        "comfort": "19.10",
-        "humidity": "74",
-        "dewPoint": "14.30",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "22",
-        "airDescription": "Nice",
-        "windSpeed": "10.08",
-        "windDirection": "283",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T09:00:00.000-04:00",
-        "localTime": "0907132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "20.70",
-        "temperatureDesc": "Cool",
-        "comfort": "20.70",
-        "humidity": "67",
-        "dewPoint": "14.40",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "10.80",
-        "windDirection": "288",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T10:00:00.000-04:00",
-        "localTime": "1007132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Cool.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "21.90",
-        "temperatureDesc": "Cool",
-        "comfort": "24.21",
-        "humidity": "64",
-        "dewPoint": "14.70",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "11.88",
-        "windDirection": "288",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T11:00:00.000-04:00",
-        "localTime": "1107132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Mild.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "23.10",
-        "temperatureDesc": "Mild",
-        "comfort": "24.75",
-        "humidity": "62",
-        "dewPoint": "15.40",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "30",
-        "airDescription": "Crisp",
-        "windSpeed": "12.60",
-        "windDirection": "286",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T12:00:00.000-04:00",
-        "localTime": "1207132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Scattered clouds. Mild.",
-        "skyInfo": "9",
-        "skyDescription": "Scattered clouds",
-        "temperature": "24.10",
-        "temperatureDesc": "Mild",
-        "comfort": "25.28",
-        "humidity": "62",
-        "dewPoint": "16.40",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "12.96",
-        "windDirection": "282",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "2",
-        "iconName": "mostly_sunny",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/2.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T13:00:00.000-04:00",
-        "localTime": "1307132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Mild.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "25.00",
-        "temperatureDesc": "Mild",
-        "comfort": "25.99",
-        "humidity": "62",
-        "dewPoint": "17.10",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "13.32",
-        "windDirection": "280",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T14:00:00.000-04:00",
-        "localTime": "1407132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Mild.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "25.30",
-        "temperatureDesc": "Mild",
-        "comfort": "26.25",
-        "humidity": "61",
-        "dewPoint": "17.10",
-        "precipitationProbability": "11",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "13.68",
-        "windDirection": "278",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T15:00:00.000-04:00",
-        "localTime": "1507132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Mild.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "25.20",
-        "temperatureDesc": "Mild",
-        "comfort": "26.15",
-        "humidity": "60",
-        "dewPoint": "16.80",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "13.68",
-        "windDirection": "277",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T16:00:00.000-04:00",
-        "localTime": "1607132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Mild.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "24.80",
-        "temperatureDesc": "Mild",
-        "comfort": "25.83",
-        "humidity": "60",
-        "dewPoint": "16.60",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "13.68",
-        "windDirection": "276",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T17:00:00.000-04:00",
-        "localTime": "1707132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Mild.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "24.10",
-        "temperatureDesc": "Mild",
-        "comfort": "25.26",
-        "humidity": "63",
-        "dewPoint": "16.70",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "12.60",
-        "windDirection": "268",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T18:00:00.000-04:00",
-        "localTime": "1807132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Mild.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "22.90",
-        "temperatureDesc": "Mild",
-        "comfort": "24.13",
-        "humidity": "70",
-        "dewPoint": "17.10",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "10.44",
-        "windDirection": "243",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T19:00:00.000-04:00",
-        "localTime": "1907132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "a mixture of sun and clouds. Cool.",
-        "skyInfo": "11",
-        "skyDescription": "a mixture of sun and clouds",
-        "temperature": "21.50",
-        "temperatureDesc": "Cool",
-        "comfort": "22.18",
-        "humidity": "79",
-        "dewPoint": "17.60",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "8.28",
-        "windDirection": "222",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T20:00:00.000-04:00",
-        "localTime": "2007132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "20.70",
-        "temperatureDesc": "Cool",
-        "comfort": "20.70",
-        "humidity": "86",
-        "dewPoint": "18.20",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.56",
-        "windDirection": "221",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T21:00:00.000-04:00",
-        "localTime": "2107132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "20.20",
-        "temperatureDesc": "Cool",
-        "comfort": "20.20",
-        "humidity": "88",
-        "dewPoint": "18.20",
-        "precipitationProbability": "11",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.20",
-        "windDirection": "225",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T22:00:00.000-04:00",
-        "localTime": "2207132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.50",
-        "temperatureDesc": "Cool",
-        "comfort": "19.50",
-        "humidity": "92",
-        "dewPoint": "18.20",
-        "precipitationProbability": "11",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.20",
-        "windDirection": "227",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "7",
-        "weekday": "Saturday",
-        "utcTime": "2019-07-13T23:00:00.000-04:00",
-        "localTime": "2307132019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.00",
-        "temperatureDesc": "Cool",
-        "comfort": "19.00",
-        "humidity": "96",
-        "dewPoint": "18.30",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.56",
-        "windDirection": "225",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T00:00:00.000-04:00",
-        "localTime": "0007142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.20",
-        "temperatureDesc": "Cool",
-        "comfort": "19.20",
-        "humidity": "95",
-        "dewPoint": "18.40",
-        "precipitationProbability": "10",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.20",
-        "windDirection": "224",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T01:00:00.000-04:00",
-        "localTime": "0107142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.80",
-        "temperatureDesc": "Cool",
-        "comfort": "19.80",
-        "humidity": "93",
-        "dewPoint": "18.60",
-        "precipitationProbability": "9",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.20",
-        "windDirection": "222",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T02:00:00.000-04:00",
-        "localTime": "0207142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.90",
-        "temperatureDesc": "Cool",
-        "comfort": "19.90",
-        "humidity": "93",
-        "dewPoint": "18.70",
-        "precipitationProbability": "8",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.20",
-        "windDirection": "217",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T03:00:00.000-04:00",
-        "localTime": "0307142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "N",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.30",
-        "temperatureDesc": "Cool",
-        "comfort": "19.30",
-        "humidity": "96",
-        "dewPoint": "18.70",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "29",
-        "airDescription": "Comfortable",
-        "windSpeed": "7.56",
-        "windDirection": "210",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "15",
-        "iconName": "night_partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/24.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T04:00:00.000-04:00",
-        "localTime": "0407142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "18.80",
-        "temperatureDesc": "Cool",
-        "comfort": "18.80",
-        "humidity": "99",
-        "dewPoint": "18.60",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.92",
-        "windDirection": "208",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T05:00:00.000-04:00",
-        "localTime": "0507142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.30",
-        "temperatureDesc": "Cool",
-        "comfort": "19.30",
-        "humidity": "99",
-        "dewPoint": "19.10",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "9.00",
-        "windDirection": "212",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T06:00:00.000-04:00",
-        "localTime": "0607142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "20.30",
-        "temperatureDesc": "Cool",
-        "comfort": "20.30",
-        "humidity": "97",
-        "dewPoint": "19.80",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "10.80",
-        "windDirection": "219",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T07:00:00.000-04:00",
-        "localTime": "0707142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "21.40",
-        "temperatureDesc": "Cool",
-        "comfort": "18.97",
-        "humidity": "94",
-        "dewPoint": "20.40",
-        "precipitationProbability": "8",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "12.24",
-        "windDirection": "226",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T08:00:00.000-04:00",
-        "localTime": "0807142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.20",
-        "temperatureDesc": "Mild",
-        "comfort": "20.47",
-        "humidity": "92",
-        "dewPoint": "20.80",
-        "precipitationProbability": "8",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "12.60",
-        "windDirection": "235",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T09:00:00.000-04:00",
-        "localTime": "0907142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.70",
-        "temperatureDesc": "Mild",
-        "comfort": "21.21",
-        "humidity": "92",
-        "dewPoint": "21.30",
-        "precipitationProbability": "8",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "13.32",
-        "windDirection": "246",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T10:00:00.000-04:00",
-        "localTime": "1007142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "23.30",
-        "temperatureDesc": "Mild",
-        "comfort": "22.33",
-        "humidity": "91",
-        "dewPoint": "21.70",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "14.04",
-        "windDirection": "257",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T11:00:00.000-04:00",
-        "localTime": "1107142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "24.20",
-        "temperatureDesc": "Mild",
-        "comfort": "24.33",
-        "humidity": "85",
-        "dewPoint": "21.50",
-        "precipitationProbability": "7",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "15.12",
-        "windDirection": "267",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T12:00:00.000-04:00",
-        "localTime": "1207142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "25.30",
-        "temperatureDesc": "Mild",
-        "comfort": "26.38",
-        "humidity": "81",
-        "dewPoint": "21.70",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "16.20",
-        "windDirection": "276",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T13:00:00.000-04:00",
-        "localTime": "1307142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "26.20",
-        "temperatureDesc": "Mild",
-        "comfort": "27.91",
-        "humidity": "78",
-        "dewPoint": "22.00",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "16.92",
-        "windDirection": "284",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T14:00:00.000-04:00",
-        "localTime": "1407142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Warm.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "26.70",
-        "temperatureDesc": "Warm",
-        "comfort": "28.71",
-        "humidity": "76",
-        "dewPoint": "22.10",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "16.20",
-        "windDirection": "288",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T15:00:00.000-04:00",
-        "localTime": "1507142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Warm.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "26.90",
-        "temperatureDesc": "Warm",
-        "comfort": "28.93",
-        "humidity": "73",
-        "dewPoint": "21.70",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "14.40",
-        "windDirection": "291",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T16:00:00.000-04:00",
-        "localTime": "1607142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Warm.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "26.90",
-        "temperatureDesc": "Warm",
-        "comfort": "28.77",
-        "humidity": "71",
-        "dewPoint": "21.30",
-        "precipitationProbability": "3",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "13.32",
-        "windDirection": "292",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T17:00:00.000-04:00",
-        "localTime": "1707142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "26.00",
-        "temperatureDesc": "Mild",
-        "comfort": "27.41",
-        "humidity": "73",
-        "dewPoint": "20.80",
-        "precipitationProbability": "3",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "12.60",
-        "windDirection": "292",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T18:00:00.000-04:00",
-        "localTime": "1807142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "24.90",
-        "temperatureDesc": "Mild",
-        "comfort": "25.80",
-        "humidity": "76",
-        "dewPoint": "20.40",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "11.88",
-        "windDirection": "291",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T19:00:00.000-04:00",
-        "localTime": "1907142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "23.50",
-        "temperatureDesc": "Mild",
-        "comfort": "23.62",
-        "humidity": "82",
-        "dewPoint": "20.20",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "10.44",
-        "windDirection": "288",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T20:00:00.000-04:00",
-        "localTime": "2007142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Mild.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "22.70",
-        "temperatureDesc": "Mild",
-        "comfort": "22.15",
-        "humidity": "85",
-        "dewPoint": "20.10",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.92",
-        "windDirection": "284",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T21:00:00.000-04:00",
-        "localTime": "2107142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "21.70",
-        "temperatureDesc": "Cool",
-        "comfort": "20.46",
-        "humidity": "88",
-        "dewPoint": "19.70",
-        "precipitationProbability": "4",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "7.20",
-        "windDirection": "278",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T22:00:00.000-04:00",
-        "localTime": "2207142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "20.60",
-        "temperatureDesc": "Cool",
-        "comfort": "20.60",
-        "humidity": "92",
-        "dewPoint": "19.30",
-        "precipitationProbability": "3",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.12",
-        "windDirection": "267",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "1",
-        "weekday": "Sunday",
-        "utcTime": "2019-07-14T23:00:00.000-04:00",
-        "localTime": "2307142019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "19.60",
-        "temperatureDesc": "Cool",
-        "comfort": "19.60",
-        "humidity": "96",
-        "dewPoint": "19.00",
-        "precipitationProbability": "5",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.12",
-        "windDirection": "252",
-        "windDesc": "West",
-        "windDescShort": "W",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-15T00:00:00.000-04:00",
-        "localTime": "0007152019",
-        "localTimeFormat": "HHMMddyyyy"
-      },
-      {
-        "daylight": "D",
-        "description": "Partly cloudy. Cool.",
-        "skyInfo": "10",
-        "skyDescription": "Partly cloudy",
-        "temperature": "18.80",
-        "temperatureDesc": "Cool",
-        "comfort": "18.80",
-        "humidity": "99",
-        "dewPoint": "18.60",
-        "precipitationProbability": "6",
-        "precipitationDesc": "",
-        "rainFall": "*",
-        "snowFall": "*",
-        "airInfo": "*",
-        "airDescription": "",
-        "windSpeed": "6.84",
-        "windDirection": "239",
-        "windDesc": "Southwest",
-        "windDescShort": "SW",
-        "visibility": "0.00",
-        "icon": "4",
-        "iconName": "partly_cloudy",
-        "iconLink": "https://weather.cit.api.here.com/static/weather/icon/3.png",
-        "dayOfWeek": "2",
-        "weekday": "Monday",
-        "utcTime": "2019-07-15T01:00:00.000-04:00",
-        "localTime": "0107152019",
-        "localTimeFormat": "HHMMddyyyy"
-      }
-    ],
-  ],
- 
+  ["weather/10001/status/hourly", getHourly],
 ];
 
 /**
