+++ src/dev/weather/forecast.js
@@ -1,0 +1,748 @@
+const getForecast = () => [
+  {
+    daylight: "D",
+    daySegment: "Afternoon",
+    description: "More clouds than sun. Cool.",
+    skyInfo: 13,
+    skyDescription: "More clouds than sun",
+    temperature: 69,
+    temperatureDesc: "Cool",
+    comfort: 68.72,
+    humidity: 68,
+    dewPoint: 58,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 10,
+    windDirection: 283,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 3,
+    beaufortDescription: "Gentle breeze",
+    visibility: 8.82,
+    icon: "6",
+    iconName: "mostly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561402800,
+  },
+  {
+    daylight: "D",
+    daySegment: "Evening",
+    description: "Broken clouds. Cool.",
+    skyInfo: 15,
+    skyDescription: "Broken clouds",
+    temperature: 64,
+    temperatureDesc: "Cool",
+    comfort: 63.5,
+    humidity: 80,
+    dewPoint: 57,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 290,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.84,
+    icon: "6",
+    iconName: "mostly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561424400,
+  },
+  {
+    daylight: "N",
+    daySegment: "Night",
+    description: "Overcast. Cool.",
+    skyInfo: 18,
+    skyDescription: "Overcast",
+    temperature: 63,
+    temperatureDesc: "Cool",
+    comfort: 62.96,
+    humidity: 81,
+    dewPoint: 57,
+    precipitationProbability: 1,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 4,
+    windDirection: 283,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.74,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561446000,
+  },
+  {
+    daylight: "D",
+    daySegment: "Morning",
+    description: "Overcast. Cool.",
+    skyInfo: 18,
+    skyDescription: "Overcast",
+    temperature: 64,
+    temperatureDesc: "Cool",
+    comfort: 64.4,
+    humidity: 78,
+    dewPoint: 57,
+    precipitationProbability: 1,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 243,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.81,
+    icon: "7",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561467600,
+  },
+  {
+    daylight: "D",
+    daySegment: "Afternoon",
+    description: "Overcast. Cool.",
+    skyInfo: 18,
+    skyDescription: "Overcast",
+    temperature: 68,
+    temperatureDesc: "Cool",
+    comfort: 67.64,
+    humidity: 68,
+    dewPoint: 57,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 9,
+    windDirection: 247,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 3,
+    beaufortDescription: "Gentle breeze",
+    visibility: 8.23,
+    icon: "7",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561489200,
+  },
+  {
+    daylight: "D",
+    daySegment: "Evening",
+    description: "Broken clouds. Cool.",
+    skyInfo: 15,
+    skyDescription: "Broken clouds",
+    temperature: 64,
+    temperatureDesc: "Cool",
+    comfort: 63.5,
+    humidity: 79,
+    dewPoint: 57,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 280,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.84,
+    icon: "6",
+    iconName: "mostly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561510800,
+  },
+  {
+    daylight: "N",
+    daySegment: "Night",
+    description: "Overcast. Cool.",
+    skyInfo: 18,
+    skyDescription: "Overcast",
+    temperature: 62,
+    temperatureDesc: "Cool",
+    comfort: 62.24,
+    humidity: 82,
+    dewPoint: 56,
+    precipitationProbability: 1,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 5,
+    windDirection: 287,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.97,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561532400,
+  },
+  {
+    daylight: "D",
+    daySegment: "Morning",
+    description: "Overcast. Cool.",
+    skyInfo: 18,
+    skyDescription: "Overcast",
+    temperature: 65,
+    temperatureDesc: "Cool",
+    comfort: 64.58,
+    humidity: 76,
+    dewPoint: 57,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 262,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.89,
+    icon: "7",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 4,
+    weekday: "Wednesday",
+    utcTime: 1561554000,
+  },
+  {
+    daylight: "D",
+    daySegment: "Afternoon",
+    description: "Broken clouds. Cool.",
+    skyInfo: 15,
+    skyDescription: "Broken clouds",
+    temperature: 68,
+    temperatureDesc: "Cool",
+    comfort: 67.82,
+    humidity: 67,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 10,
+    windDirection: 260,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 3,
+    beaufortDescription: "Gentle breeze",
+    visibility: 7.75,
+    icon: "6",
+    iconName: "mostly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/17.png",
+    dayOfWeek: 4,
+    weekday: "Wednesday",
+    utcTime: 1561575600,
+  },
+  {
+    daylight: "N",
+    daySegment: "Evening",
+    description: "Overcast. Cool.",
+    skyInfo: 18,
+    skyDescription: "Overcast",
+    temperature: 63,
+    temperatureDesc: "Cool",
+    comfort: 62.78,
+    humidity: 80,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 290,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 7.95,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 4,
+    weekday: "Wednesday",
+    utcTime: 1561597200,
+  },
+  {
+    daylight: "N",
+    daySegment: "Night",
+    description: "Cloudy. Cool.",
+    skyInfo: 17,
+    skyDescription: "Cloudy",
+    temperature: 60,
+    temperatureDesc: "Cool",
+    comfort: 60.44,
+    humidity: 84,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 4,
+    windDirection: 293,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    beaufortScale: 1,
+    beaufortDescription: "Light air",
+    visibility: 0,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 4,
+    weekday: "Wednesday",
+    utcTime: 1561618800,
+  },
+  {
+    daylight: "D",
+    daySegment: "Morning",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 65,
+    temperatureDesc: "Cool",
+    comfort: 65.12,
+    humidity: 73,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 237,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 5,
+    weekday: "Thursday",
+    utcTime: 1561640400,
+  },
+  {
+    daylight: "D",
+    daySegment: "Afternoon",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 71,
+    temperatureDesc: "Cool",
+    comfort: 75.74,
+    humidity: 62,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 10,
+    windDirection: 235,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 3,
+    beaufortDescription: "Gentle breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 5,
+    weekday: "Thursday",
+    utcTime: 1561662000,
+  },
+  {
+    daylight: "N",
+    daySegment: "Evening",
+    description: "Cloudy. Cool.",
+    skyInfo: 17,
+    skyDescription: "Cloudy",
+    temperature: 65,
+    temperatureDesc: "Cool",
+    comfort: 65.48,
+    humidity: 78,
+    dewPoint: 58,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 6,
+    windDirection: 286,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 5,
+    weekday: "Thursday",
+    utcTime: 1561683600,
+  },
+  {
+    daylight: "N",
+    daySegment: "Night",
+    description: "Cloudy. Cool.",
+    skyInfo: 17,
+    skyDescription: "Cloudy",
+    temperature: 62,
+    temperatureDesc: "Cool",
+    comfort: 62.78,
+    humidity: 81,
+    dewPoint: 57,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 5,
+    windDirection: 247,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 5,
+    weekday: "Thursday",
+    utcTime: 1561705200,
+  },
+  {
+    daylight: "D",
+    daySegment: "Morning",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 68,
+    temperatureDesc: "Cool",
+    comfort: 67.82,
+    humidity: 67,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 226,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 6,
+    weekday: "Friday",
+    utcTime: 1561726800,
+  },
+  {
+    daylight: "D",
+    daySegment: "Afternoon",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 71,
+    temperatureDesc: "Cool",
+    comfort: 76.23,
+    humidity: 59,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 10,
+    windDirection: 255,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 3,
+    beaufortDescription: "Gentle breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 6,
+    weekday: "Friday",
+    utcTime: 1561748400,
+  },
+  {
+    daylight: "N",
+    daySegment: "Evening",
+    description: "Partly cloudy. Cool.",
+    skyInfo: 10,
+    skyDescription: "Partly cloudy",
+    temperature: 64,
+    temperatureDesc: "Cool",
+    comfort: 64.4,
+    humidity: 77,
+    dewPoint: 57,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 6,
+    windDirection: 260,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "15",
+    iconName: "night_partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 6,
+    weekday: "Friday",
+    utcTime: 1561770000,
+  },
+  {
+    daylight: "N",
+    daySegment: "Night",
+    description: "Partly cloudy. Cool.",
+    skyInfo: 10,
+    skyDescription: "Partly cloudy",
+    temperature: 61,
+    temperatureDesc: "Cool",
+    comfort: 61.88,
+    humidity: 83,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 4,
+    windDirection: 203,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "15",
+    iconName: "night_partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 6,
+    weekday: "Friday",
+    utcTime: 1561791600,
+  },
+  {
+    daylight: "D",
+    daySegment: "Morning",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 69,
+    temperatureDesc: "Cool",
+    comfort: 68.72,
+    humidity: 66,
+    dewPoint: 57,
+    precipitationProbability: 1,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "*",
+    airDescription: "",
+    windSpeed: 6,
+    windDirection: 277,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 7,
+    weekday: "Saturday",
+    utcTime: 1561813200,
+  },
+  {
+    daylight: "D",
+    daySegment: "Afternoon",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 74,
+    temperatureDesc: "Cool",
+    comfort: 76.87,
+    humidity: 56,
+    dewPoint: 56,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 10,
+    windDirection: 272,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 3,
+    beaufortDescription: "Gentle breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 7,
+    weekday: "Saturday",
+    utcTime: 1561834800,
+  },
+  {
+    daylight: "N",
+    daySegment: "Evening",
+    description: "Partly cloudy. Cool.",
+    skyInfo: 10,
+    skyDescription: "Partly cloudy",
+    temperature: 66,
+    temperatureDesc: "Cool",
+    comfort: 66.38,
+    humidity: 76,
+    dewPoint: 59,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 6,
+    windDirection: 280,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "15",
+    iconName: "night_partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 7,
+    weekday: "Saturday",
+    utcTime: 1561856400,
+  },
+  {
+    daylight: "N",
+    daySegment: "Night",
+    description: "Cloudy. Cool.",
+    skyInfo: 17,
+    skyDescription: "Cloudy",
+    temperature: 63,
+    temperatureDesc: "Cool",
+    comfort: 64.04,
+    humidity: 80,
+    dewPoint: 58,
+    precipitationProbability: 2,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 5,
+    windDirection: 281,
+    windDesc: "West",
+    windDescShort: "W",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "17",
+    iconName: "cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/24.png",
+    dayOfWeek: 7,
+    weekday: "Saturday",
+    utcTime: 1561878000,
+  },
+  {
+    daylight: "D",
+    daySegment: "Morning",
+    description: "a mixture of sun and clouds. Cool.",
+    skyInfo: 11,
+    skyDescription: "a mixture of sun and clouds",
+    temperature: 70,
+    temperatureDesc: "Cool",
+    comfort: 75.58,
+    humidity: 62,
+    dewPoint: 57,
+    precipitationProbability: 4,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 6,
+    windDirection: 315,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    beaufortScale: 2,
+    beaufortDescription: "Light breeze",
+    visibility: 0,
+    icon: "4",
+    iconName: "partly_cloudy",
+    iconLink: "https://weather.api.here.com/static/weather/icon/3.png",
+    dayOfWeek: 1,
+    weekday: "Sunday",
+    utcTime: 1561899600,
+  },
+];
+
+export { getForecast };
