+++ src/dev/weather/hourly.js
@@ -1,0 +1,696 @@
+import { randomTemp, randomPositive } from "../utils";
+
+const getHourly = () => [
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: randomPositive(3),
+    skyDescription: "Sunny",
+    temperature: randomTemp(),
+    temperatureDesc: "Hot",
+    comfort: randomTemp(),
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: randomPositive(20),
+    windDirection: randomPositive(360),
+    windDesc: "South",
+    windDescShort: "S",
+    visibility: randomPositive(40),
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561399200,
+    localTime: "1106242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Mostly sunny. Cool.",
+    skyInfo: randomPositive(3),
+    skyDescription: "Mostly sunny",
+    temperature: randomTemp(),
+    temperatureDesc: "Cool",
+    comfort: randomTemp(),
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 8,
+    windDirection: 152,
+    windDesc: "North",
+    windDescShort: "N",
+    visibility: 33.97,
+    icon: "1",
+    iconName: "mostly_sunny",
+    iconLink: "https://weather.cit.api.here.com/static/weather/icon/2.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561402800,
+    localTime: "1206242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 104,
+    temperatureDesc: "Hot",
+    comfort: 98.4,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 9,
+    windDirection: 149,
+    windDesc: "Southeast",
+    windDescShort: "SE",
+    visibility: 36.37,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561406400,
+    localTime: "1306242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 105,
+    temperatureDesc: "Hot",
+    comfort: 99.14,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 10,
+    windDirection: 148,
+    windDesc: "Southeast",
+    windDescShort: "SE",
+    visibility: 38.44,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561410000,
+    localTime: "1406242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 106,
+    temperatureDesc: "Hot",
+    comfort: 100,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 9,
+    windDirection: 152,
+    windDesc: "Southeast",
+    windDescShort: "SE",
+    visibility: 39.18,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561413600,
+    localTime: "1506242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 106,
+    temperatureDesc: "Hot",
+    comfort: 99.82,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 8,
+    windDirection: 170,
+    windDesc: "South",
+    windDescShort: "S",
+    visibility: 39.87,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561417200,
+    localTime: "1606242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 105,
+    temperatureDesc: "Hot",
+    comfort: 99.46,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 8,
+    windDirection: 226,
+    windDesc: "Southwest",
+    windDescShort: "SW",
+    visibility: 40.53,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561420800,
+    localTime: "1706242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 103,
+    temperatureDesc: "Hot",
+    comfort: 98.29,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 9,
+    windDirection: 275,
+    windDesc: "West",
+    windDescShort: "W",
+    visibility: 37.8,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561424400,
+    localTime: "1806242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Hot.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 99,
+    temperatureDesc: "Hot",
+    comfort: 94.53,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 10,
+    windDirection: 295,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 34.37,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561428000,
+    localTime: "1906242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Hot.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 96,
+    temperatureDesc: "Hot",
+    comfort: 91.87,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "28",
+    airDescription: "Dry",
+    windSpeed: 11,
+    windDirection: 305,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 30.06,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561431600,
+    localTime: "2006242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Pleasantly warm.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 89,
+    temperatureDesc: "Pleasantly warm",
+    comfort: 86.18,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 11,
+    windDirection: 316,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 27.02,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561435200,
+    localTime: "2106242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 86,
+    temperatureDesc: "Mild",
+    comfort: 84.09,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 11,
+    windDirection: 321,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 23.52,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561438800,
+    localTime: "2206242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 85,
+    temperatureDesc: "Mild",
+    comfort: 83.16,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 10,
+    windDirection: 323,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 19.49,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 2,
+    weekday: "Monday",
+    utcTime: 1561442400,
+    localTime: "2306242019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 82,
+    temperatureDesc: "Mild",
+    comfort: 81.05,
+    humidity: randomPositive(100),
+    dewPoint: randomPositive(100),
+    precipitationProbability: randomPositive(100),
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 9,
+    windDirection: 323,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 17.27,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561446000,
+    localTime: "0006252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 80,
+    temperatureDesc: "Mild",
+    comfort: 79.93,
+    humidity: 39,
+    dewPoint: 53,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 8,
+    windDirection: 323,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 14.87,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561449600,
+    localTime: "0106252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 79,
+    temperatureDesc: "Mild",
+    comfort: 79.05,
+    humidity: 40,
+    dewPoint: 52,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 7,
+    windDirection: 324,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 12.3,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561453200,
+    localTime: "0206252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 76,
+    temperatureDesc: "Mild",
+    comfort: 78.08,
+    humidity: 42,
+    dewPoint: 52,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 7,
+    windDirection: 323,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 9.72,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561456800,
+    localTime: "0306252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "N",
+    description: "Clear. Mild.",
+    skyInfo: 2,
+    skyDescription: "Clear",
+    temperature: 76,
+    temperatureDesc: "Mild",
+    comfort: 77.74,
+    humidity: 42,
+    dewPoint: 51,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 7,
+    windDirection: 322,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 9.16,
+    icon: "13",
+    iconName: "night_clear",
+    iconLink: "https://weather.api.here.com/static/weather/icon/16.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561460400,
+    localTime: "0406252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Cool.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 75,
+    temperatureDesc: "Cool",
+    comfort: 77.5,
+    humidity: 42,
+    dewPoint: 50,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "20",
+    airDescription: "Beautiful",
+    windSpeed: 7,
+    windDirection: 321,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 9.72,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561464000,
+    localTime: "0506252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Mild.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 75,
+    temperatureDesc: "Mild",
+    comfort: 77.58,
+    humidity: 42,
+    dewPoint: 50,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "20",
+    airDescription: "Beautiful",
+    windSpeed: 7,
+    windDirection: 323,
+    windDesc: "Northwest",
+    windDescShort: "NW",
+    visibility: 9.28,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561467600,
+    localTime: "0606252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Mild.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 77,
+    temperatureDesc: "Mild",
+    comfort: 78.35,
+    humidity: 39,
+    dewPoint: 51,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 6,
+    windDirection: 32,
+    windDesc: "Northeast",
+    windDescShort: "NE",
+    visibility: 9.73,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561471200,
+    localTime: "0706252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Mild.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 82,
+    temperatureDesc: "Mild",
+    comfort: 80.76,
+    humidity: 34,
+    dewPoint: 51,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 6,
+    windDirection: 134,
+    windDesc: "Southeast",
+    windDescShort: "SE",
+    visibility: 13.28,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561474800,
+    localTime: "0806252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+  {
+    daylight: "D",
+    description: "Sunny. Mild.",
+    skyInfo: 1,
+    skyDescription: "Sunny",
+    temperature: 86,
+    temperatureDesc: "Mild",
+    comfort: 83.35,
+    humidity: 28,
+    dewPoint: 49,
+    precipitationProbability: 0,
+    precipitationDesc: "",
+    rainFall: "*",
+    snowFall: "*",
+    airInfo: "29",
+    airDescription: "Comfortable",
+    windSpeed: 7,
+    windDirection: 145,
+    windDesc: "Southeast",
+    windDescShort: "SE",
+    visibility: 16.89,
+    icon: "1",
+    iconName: "sunny",
+    iconLink: "https://weather.api.here.com/static/weather/icon/1.png",
+    dayOfWeek: 3,
+    weekday: "Tuesday",
+    utcTime: 1561478400,
+    localTime: "0906252019",
+    localTimeFormat: "HHMMddyyyy",
+  },
+];
+
+export { getHourly };
